#132 checkAndFilterContingencies now adds to a vector timeOfEvent is a float now timeStart_ and durationSimu_ are float now too

+++ sources/Inputs/include/Configuration.h
@@ -121,7 +121,7 @@ class Configuration {
    *
    * @returns the parameter value
    */
-  double getTimeOfEvent() const {
+  float getTimeOfEvent() const {
     return timeOfEvent_;
   }
